Split OrderItemService to interface and implementation class
